@@ -10,21 +10,12 @@ class NamedError extends Error {
     return `${this.name}: ${this.message}`;
   }
 }
-class EncryptionError extends NamedError {}
-class EncryptionJsonError extends EncryptionError {}
-class DecryptionError extends NamedError {}
-class DecryptionIntegrityError extends DecryptionError {}
-class DecryptionSignatureError extends DecryptionError {}
-class DecryptionTimeTravelError extends DecryptionError {}
-class DecryptionExpiredError extends DecryptionError {}
-class DecryptionGunzipError extends DecryptionError {}
-class DecryptionJsonError extends DecryptionError {}
-module.exports.EncryptionError = EncryptionError;
-module.exports.EncryptionJsonError = EncryptionJsonError;
-module.exports.DecryptionError = DecryptionError;
-module.exports.DecryptionIntegrityError = DecryptionIntegrityError;
-module.exports.DecryptionSignatureError = DecryptionSignatureError;
-module.exports.DecryptionTimeTravelError = DecryptionTimeTravelError;
-module.exports.DecryptionExpiredError = DecryptionExpiredError;
-module.exports.DecryptionGunzipError = DecryptionGunzipError;
-module.exports.DecryptionJsonError = DecryptionJsonError;
+export class EncryptionError extends NamedError {}
+export class EncryptionJsonError extends EncryptionError {}
+export class DecryptionError extends NamedError {}
+export class DecryptionIntegrityError extends DecryptionError {}
+export class DecryptionSignatureError extends DecryptionError {}
+export class DecryptionTimeTravelError extends DecryptionError {}
+export class DecryptionExpiredError extends DecryptionError {}
+export class DecryptionGunzipError extends DecryptionError {}
+export class DecryptionJsonError extends DecryptionError {}
