@@ -1,6 +1,8 @@
-module.exports.constants = require('./constants');
-module.exports.errors = require('./errors');
-module.exports.toUrlSafeBase64 = require('./url-safe-base64').encode;
-module.exports.fromUrlSafeBase64 = require('./url-safe-base64').decode;
-module.exports.Crypter = require('./crypter').Crypter;
-module.exports.Chester = require('./chester').Chester;
+export * as constants from './constants.js';
+export * as errors from './errors.js';
+export {
+  encode as toUrlSafeBase64,
+  decode as fromUrlSafeBase64
+} from './url-safe-base64.js';
+export * from './crypter.js';
+export * from './chester.js';
